Move component to src/examples and display conditionally.
