+++ src/App.jsx
@@ -22,7 +22,7 @@ import Metadata from "./Metadata";
 import NodeInfo from "./NodeInfo";
 import Transfer from "./Transfer";
 import Upgrade from "./Upgrade";
-import TemplateModule from "./TemplateModule";
+import TemplateModule from "./examples/TemplateModule";
 import "semantic-ui-css/semantic.min.css";
 
 export default function App() {
@@ -156,12 +156,14 @@ export default function App() {
           <Grid.Row>
             <Transfer api={api} accountPair={accountPair} />
             <Upgrade api={api} accountPair={accountPair} />
-            <TemplateModule api={api} accountPair={accountPair} />
           </Grid.Row>
           <Grid.Row>
             <Extrinsics api={api} accountPair={accountPair} />
             <ChainState api={api} />
             <Events api={api} />
+          </Grid.Row>
+          <Grid.Row>
+            { api.query.templateModule && <TemplateModule api={api} accountPair={accountPair} /> }
           </Grid.Row>
         </Grid>
         {/* These components don't render elements. */}
